pyproject.toml
[project]
name = "youcanbechef"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiokafka>=0.12.0",
    "aiosmtplib>=5.0.0",
    "alembic>=1.17.1",
    "asyncpg>=0.30.0",
    "black>=25.11.0",
    "email-validator>=2.3.0",
    "fastapi>=0.121.1",
    "httpx>=0.28.1",
    "pwdlib[argon2]>=0.3.0",
    "pydantic-settings>=2.12.0",
    "pydantic[email]>=2.12.4",
    "pytest>=9.0.0",
    "pytest-asyncio>=1.3.0",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "redis>=7.0.1",
    "ruff>=0.14.4",
    "sqlalchemy[asyncio]>=2.0.44",
    "uvicorn[standard]>=0.38.0",
]
